+++ test/clj/rems/db/test_entitlements.clj
@@ -1,12 +1,12 @@
 (ns ^:integration rems.db.test-entitlements
-  (:require [cheshire.core :as cheshire]
-            [clj-time.core :as time]
+  (:require [clj-time.core :as time]
             [clojure.test :refer :all]
             [rems.db.applications :as applications]
             [rems.db.core :as db]
             [rems.db.entitlements :as entitlements]
             [rems.db.test-data :as test-data]
             [rems.db.testing :refer [test-db-fixture rollback-db-fixture test-data-fixture]]
+            [rems.json :as json]
             [rems.testing-util :refer [suppress-logging]]
             [stub-http.core :as stub]))
 
@@ -22,51 +22,40 @@
    {:resid "res2" :catappid 12 :userid "user2" :start (time/date-time 2002 10 11) :mail "user2@tes.t"}])
 
 (def +expected-payload+
-  [{"resource" "res1" "application" 11 "user" "user1" "mail" "user1@tes.t"}
-   {"resource" "res2" "application" 12 "user" "user2" "mail" "user2@tes.t"}])
+  [{:resource "res1" :application 11 :user "user1" :mail "user1@tes.t"}
+   {:resource "res2" :application 12 :user "user2" :mail "user2@tes.t"}])
 
 (defn run-with-server
-  "Run callback with a mock entitlements http server set up.
-   Return sequence of data received by mock server."
   [endpoint-spec callback]
   (with-open [server (stub/start! {"/entitlements" endpoint-spec})]
     (with-redefs [rems.config/env (assoc rems.config/env
                                          :entitlements-target {:add (str (:uri server) "/entitlements")})]
-      (callback)
-      (for [r (stub/recorded-requests server)]
-        (cheshire/parse-string (get-in r [:body "postData"]))))))
+      (callback server))))
 
-(deftest test-post-entitlements
-  (let [log (atom [])]
-    (with-redefs [db/log-entitlement-post! #(swap! log conj %)]
-      (testing "ok"
-        (is (= [+expected-payload+]
-               (run-with-server {:status 200}
-                                #(#'entitlements/post-entitlements :add +entitlements+))))
-        (let [[{payload :payload status :status}] @log]
-          (is (= 200 status))
-          (is (= +expected-payload+ (cheshire/parse-string payload))))
-        (reset! log []))
-      (testing "not found"
-        (run-with-server {:status 404}
-                         #(#'entitlements/post-entitlements :add +entitlements+))
-        (let [[{payload :payload status :status}] @log]
-          (is (= 404 status))
-          (is (= +expected-payload+ (cheshire/parse-string payload))))
-        (reset! log []))
-      (testing "timeout"
-        (run-with-server {:status 200 :delay 5000} ;; timeout of 2500 in code
-                         #(#'entitlements/post-entitlements :add +entitlements+))
-        (let [[{payload :payload status :status}] @log]
-          (is (= "exception" status))
-          (is (= +expected-payload+ (cheshire/parse-string payload)))))
-      (testing "no server"
-        (with-redefs [rems.config/env (assoc rems.config/env
-                                             :entitlements-target "http://invalid/entitlements")]
-          (#'entitlements/post-entitlements :add +entitlements+)
-          (let [[{payload :payload status :status}] @log]
-            (is (= "exception" status))
-            (is (= +expected-payload+ (cheshire/parse-string payload)))))))))
+(deftest test-post-entitlements!
+  (testing "ok"
+    (run-with-server
+     {:status 200}
+     (fn [server]
+       (is (nil? (#'entitlements/post-entitlements! {:action :add :entitlements +entitlements+})))
+       (is (= [+expected-payload+] (for [r (stub/recorded-requests server)]
+                                     (json/parse-string (get-in r [:body "postData"]))))))))
+  (testing "not-found"
+    (run-with-server
+     {:status 404}
+     (fn [_]
+       (is (= "failed: 404" (#'entitlements/post-entitlements! {:action :add :entitlements +entitlements+}))))))
+  (testing "timeout"
+    (run-with-server
+     {:status 200 :delay 5000} ;; timeout of 2500 in code
+     (fn [_]
+       (is (= "failed: exception" (#'entitlements/post-entitlements! {:action :add :entitlements +entitlements+}))))))
+  (testing "invalid url"
+    (with-redefs [rems.config/env (assoc rems.config/env
+                                         :entitlements-target {:add "http://invalid/entitlements"})]
+      (is (= "failed: exception" (#'entitlements/post-entitlements! {:action :add :entitlements +entitlements+})))))
+  (testing "no server configured"
+    (is (nil? (#'entitlements/post-entitlements! {:action :add :entitlements +entitlements+})))))
 
 (defmacro with-stub-server [sym & body]
   `(let [~sym (stub/start! {"/add" {:status 200}
@@ -75,6 +64,12 @@
                                           :entitlements-target {:add (str (:uri ~sym) "/add")
                                                                 :remove (str (:uri ~sym) "/remove")})]
        ~@body)))
+
+(defn- get-requests [server]
+  (doall
+   (for [req (stub/recorded-requests server)]
+     {:path (:path req)
+      :body (json/parse-string (get-in req [:body "postData"]))})))
 
 (deftest test-entitlement-granting
   (let [applicant "bob"
@@ -103,6 +98,8 @@
     (test-data/create-user! {:eppn member :mail "e.l@s.a" :commonName "Elsa"})
     (test-data/create-user! {:eppn admin :mail "o.w@n.er" :commonName "Owner"})
 
+    (entitlements/process-outbox!) ;; empty outbox from pending posts
+
     (let [app-id (test-data/create-application! {:actor applicant :catalogue-item-ids [item1 item2]})]
       (testing "submitted application should not yet cause entitlements"
         (with-stub-server server
@@ -117,6 +114,8 @@
                                :application-id app-id
                                :actor admin
                                :member {:userid member}})
+
+          (entitlements/process-outbox!)
 
           (is (empty? (db/get-entitlements {:application app-id})))
           (is (empty? (stub/recorded-requests server))))
@@ -135,21 +134,15 @@
                     member #{lic-id1}}
                    (:application/accepted-licenses (applications/get-unrestricted-application app-id))))
 
-            (is (= 2 (count (stub/recorded-requests server))))
+            (entitlements/process-outbox!)
+
             (testing "db"
               (is (= [[applicant "resource1"] [applicant "resource2"]]
                      (map (juxt :userid :resid) (db/get-entitlements {:application app-id})))))
             (testing "POST"
-              (let [data (take 2 (stub/recorded-requests server))
-                    targets (map :path data)
-                    bodies (->> data
-                                (map #(get-in % [:body "postData"]))
-                                (map #(cheshire/parse-string % keyword))
-                                (apply concat))]
-                (is (every? #{"/add"} targets))
-                (is (= [{:resource "resource1" :application app-id :user "bob" :mail "b@o.b"}
-                        {:resource "resource2" :application app-id :user "bob" :mail "b@o.b"}]
-                       bodies)))))))
+              (is (= [{:path "/add" :body [{:application app-id :mail "b@o.b" :resource "resource1" :user "bob"}]}
+                      {:path "/add" :body [{:application app-id :mail "b@o.b" :resource "resource2" :user "bob"}]}]
+                     (get-requests server)))))))
 
       (testing "approved application, more accepted licenses generates more entitlements"
         (with-stub-server server
@@ -158,22 +151,16 @@
                                :actor member
                                :accepted-licenses [lic-id1 lic-id2]}) ; now accept all licenses
 
-          (is (= 2 (count (stub/recorded-requests server))))
+          (entitlements/process-outbox!)
+
           (testing "db"
             (is (= [[applicant "resource1"] [applicant "resource2"]
                     [member "resource1"] [member "resource2"]]
                    (map (juxt :userid :resid) (db/get-entitlements {:application app-id})))))
           (testing "POST"
-            (let [data (stub/recorded-requests server)
-                  targets (map :path data)
-                  bodies (->> data
-                              (map #(get-in % [:body "postData"]))
-                              (map #(cheshire/parse-string % keyword))
-                              (apply concat))]
-              (is (every? #{"/add"} targets))
-              (is (= [{:resource "resource1" :application app-id :user "elsa" :mail "e.l@s.a"}
-                      {:resource "resource2" :application app-id :user "elsa" :mail "e.l@s.a"}]
-                     bodies))))))
+              (is (= [{:path "/add" :body [{:resource "resource1" :application app-id :user "elsa" :mail "e.l@s.a"}]}
+                      {:path "/add" :body [{:resource "resource2" :application app-id :user "elsa" :mail "e.l@s.a"}]}]
+                     (get-requests server))))))
 
       (testing "removing a member ends entitlements"
         (with-stub-server server
@@ -183,21 +170,15 @@
                                :member {:userid member}
                                :comment "Left team"})
 
-          (is (= 2 (count (stub/recorded-requests server))))
+          (entitlements/process-outbox!)
+
           (testing "db"
             (is (= [[applicant "resource1"] [applicant "resource2"]]
                    (map (juxt :userid :resid) (db/get-entitlements {:application app-id :is-active? true})))))
           (testing "POST"
-            (let [data (stub/recorded-requests server)
-                  targets (map :path data)
-                  bodies (->> data
-                              (map #(get-in % [:body "postData"]))
-                              (map #(cheshire/parse-string % keyword))
-                              (apply concat))]
-              (is (every? #{"/remove"} targets))
-              (is (= [{:resource "resource1" :application app-id :user "elsa" :mail "e.l@s.a"}
-                      {:resource "resource2" :application app-id :user "elsa" :mail "e.l@s.a"}]
-                     bodies))))))
+            (is (= [{:path "/remove" :body [{:resource "resource1" :application app-id :user "elsa" :mail "e.l@s.a"}]}
+                    {:path "/remove" :body [{:resource "resource2" :application app-id :user "elsa" :mail "e.l@s.a"}]}]
+                   (get-requests server))))))
 
       (testing "changing resources changes entitlements"
         (with-stub-server server
@@ -207,16 +188,15 @@
                                :catalogue-item-ids [item1 item3]
                                :comment "Removed second resource, added third resource"})
 
-          (is (= 2 (count (stub/recorded-requests server))))
+          (entitlements/process-outbox!)
+
           (testing "db"
             (is (= [[applicant "resource1"] [applicant "resource3"]]
                    (map (juxt :userid :resid) (db/get-entitlements {:application app-id :is-active? true})))))
           (testing "POST"
-            (let [data (stub/recorded-requests server)
-                  data (map #(assoc % :body (cheshire/parse-string (get-in % [:body "postData"]) keyword)) data)]
-              (is (= [{:path "/add" :body [{:resource "resource3" :application app-id :user "bob" :mail "b@o.b"}]}
-                      {:path "/remove" :body [{:resource "resource2" :application app-id :user "bob" :mail "b@o.b"}]}]
-                     (map #(select-keys % [:path :body]) data)))))))
+            (is (= [{:path "/add" :body [{:resource "resource3" :application app-id :user "bob" :mail "b@o.b"}]}
+                    {:path "/remove" :body [{:resource "resource2" :application app-id :user "bob" :mail "b@o.b"}]}]
+                   (get-requests server))))))
 
       (testing "closed application should end entitlements"
         (with-stub-server server
@@ -225,15 +205,14 @@
                                :actor admin
                                :comment "Finished"})
 
-          (is (= 2 (count (stub/recorded-requests server))))
+          (entitlements/process-outbox!)
+
           (testing "db"
             (is (= [] (db/get-entitlements {:application app-id :is-active? true}))))
           (testing "POST"
-            (let [data (stub/recorded-requests server)
-                  data (map #(assoc % :body (cheshire/parse-string (get-in % [:body "postData"]) keyword)) data)]
-              (is (= [{:path "/remove" :body [{:resource "resource1" :application app-id :user "bob" :mail "b@o.b"}]}
-                      {:path "/remove" :body [{:resource "resource3" :application app-id :user "bob" :mail "b@o.b"}]}]
-                     (map #(select-keys % [:path :body]) data))))))))
+            (is (= [{:path "/remove" :body [{:resource "resource1" :application app-id :user "bob" :mail "b@o.b"}]}
+                    {:path "/remove" :body [{:resource "resource3" :application app-id :user "bob" :mail "b@o.b"}]}]
+                   (get-requests server)))))))
 
     (let [app-id (test-data/create-application! {:actor applicant :catalogue-item-ids [item1]})]
       (test-data/command! {:type :application.command/accept-licenses
@@ -248,6 +227,8 @@
                            :actor admin
                            :comment ""})
 
+      (entitlements/process-outbox!)
+
       (testing "revoked application should end entitlements"
         (with-stub-server server
           (test-data/command! {:type :application.command/revoke
@@ -255,11 +236,10 @@
                                :actor admin
                                :comment "Banned"})
 
-          (is (= 1 (count (stub/recorded-requests server))))
+          (entitlements/process-outbox!)
+
           (testing "db"
             (is (= [] (db/get-entitlements {:application app-id :is-active? true}))))
           (testing "POST"
-            (let [data (stub/recorded-requests server)
-                  data (map #(assoc % :body (cheshire/parse-string (get-in % [:body "postData"]) keyword)) data)]
-              (is (= [{:path "/remove" :body [{:resource "resource1" :application app-id :user "bob" :mail "b@o.b"}]}]
-                     (map #(select-keys % [:path :body]) data))))))))))
+            (is (= [{:path "/remove" :body [{:resource "resource1" :application app-id :user "bob" :mail "b@o.b"}]}]
+                   (get-requests server)))))))))